user can now able to add song

+++ backend/src/controllers/song.controller.js
@@ -1,6 +1,6 @@
 const songModel = require("../models/song.model");
-
-
+const { uploadSong, uploadImage } = require("../services/storage.service");
+const { v4: uuidv4 } = require("uuid");
 
 // -------- Controller to get all songs --------
 async function getSongs(req, res) {
@@ -14,20 +14,33 @@ async function getSongs(req, res) {
 
 // -------- Controller to add a new song --------
 async function addSong(req, res) {
+  if (req.user.type !== "artist") {
+    return res.status(500).json({
+      message: "only artist can add songs.",
+    });
+  }
   try {
-    const { title, artist, album, url, coverImage, lyrics, category } =
-      req.body;
-    if (!title || !artist || !url || !category) {
+    const { title, artist, album, lyrics, category } = req.body;
+    const song = req.files.song[0];
+    const coverImage = req.files.coverImage[0];
+
+    if (!title || !artist || !song || !category) {
       return res
         .status(400)
         .json({ message: "Title, artist, URL, and category are required" });
     }
+
+    const [uploadedSong, uploadedImage] = await Promise.all([
+      uploadSong(song.buffer, `${uuidv4()}`),
+      uploadImage(coverImage.buffer, `${uuidv4()}`),
+    ]);
+
     const newSong = new songModel({
       title,
       artist,
       album,
-      url,
-      coverImage,
+      url: uploadedSong.url,
+      coverImage: uploadedImage.url,
       lyrics,
       category,
     });
@@ -40,6 +53,11 @@ async function addSong(req, res) {
 
 // -------- Controller to delete a song --------
 async function deleteSong(req, res) {
+  if (req.user.type !== "artist") {
+    return res.status(500).json({
+      message: "only artist can add songs.",
+    });
+  }
   try {
     const { id } = req.params;
     const deletedSong = await songModel.findByIdAndDelete(id);
